+++ CHANGES.md
@@ -4,6 +4,8 @@
 - Allow %{read-lines} to introduce dynamic dependencies like %{read}. (#5440,
   @anmonteiro)
 
+- Look up `gmake` before `make` (#5474, fixes #5470, @rgrinberg)
+
 3.0.2 (17/02/2022)
 ------------------